Completing the testing (Business Layer)

+++ src/main/java/com/abde/customer/CustomerController.java
@@ -37,7 +37,7 @@ public class CustomerController {
     }
 
     @DeleteMapping("{customerId}")
-    public void deleteCustomer(@PathVariable("customerId") Integer customerId) {
+    public void deleteCustomer(@PathVariable("customerId") Long customerId) {
         customerService.deleteCustomer(customerId);
     }
 }
